non-saturating small cfloats using a block arithmetic abstraction

+++ include/universal/verification/test_suite_random.hpp
@@ -205,7 +205,7 @@ namespace sw::universal {
 	// Basic design is that we generate nrOfRandom posit values and store them in an operand array.
 	// We will then execute the binary operator nrOfRandom combinations.
 	template<typename TestType>
-	int VerifyBinaryOperatorThroughRandoms(bool bReportIndividualTestCases, int opcode, uint32_t nrOfRandoms) {
+	int VerifyBinaryOperatorThroughRandoms(bool bReportIndividualTestCases, int opcode, size_t nrOfRandoms) {
 		std::cerr << typeid(TestType).name() << " : ";
 
 		std::string operation_string;
@@ -277,7 +277,7 @@ namespace sw::universal {
 	// We will then execute the binary operator nrOfRandom combinations.
 	// provide 		double dminpos = double(minpos<nbits, es>(pminpos));
 	template<typename TestType>
-	int VerifyUnaryOperatorThroughRandoms(bool bReportIndividualTestCases, int opcode, uint32_t nrOfRandoms, double dminpos) {
+	int VerifyUnaryOperatorThroughRandoms(bool bReportIndividualTestCases, int opcode, size_t nrOfRandoms, double dminpos) {
 		std::string operation_string;
 		bool sqrtOperator = false;  // we need to filter negative values from the randoms
 		switch (opcode) {
